feat: add all bundled otel config to the observe-agent binary instead of relying on files on disk

+++ scripts/install.ps1
@@ -24,7 +24,6 @@ $temp_dir="C:\temp"
 New-Item -ItemType Directory -Force -Path $temp_dir
 New-Item -ItemType Directory -Force -Path $observeagent_install_dir
 New-Item -ItemType Directory -Force -Path $observeagent_install_dir\config
-New-Item -ItemType Directory -Force -Path $observeagent_install_dir\connections
 New-Item -ItemType Directory -Force -Path $program_data_filestorage
 
 if ($PSBoundParameters.ContainsKey('zip_dir')){
@@ -45,7 +44,10 @@ Copy-Item -Force -Path $temp_dir\observe-agent_Windows_x86_64\observe-agent.exe 
 if (Test-Path $temp_dir\observe-agent_Windows_x86_64\otel-collector.yaml) {
     Copy-Item -Force -Path $temp_dir\observe-agent_Windows_x86_64\otel-collector.yaml -Destination $observeagent_install_dir\config\otel-collector.yaml
 }
-Copy-Item -Force -Path $temp_dir\observe-agent_Windows_x86_64\connections\* -Destination $observeagent_install_dir\connections -Recurse
+if (Test-Path $temp_dir\observe-agent_Windows_x86_64\connections) {
+    New-Item -ItemType Directory -Force -Path $observeagent_install_dir\connections
+    Copy-Item -Force -Path $temp_dir\observe-agent_Windows_x86_64\connections\* -Destination $observeagent_install_dir\connections -Recurse
+}
 
 # If there's already an observe-agent.yaml we don't copy the template from the downloaded installation package and leave existing observe-agent.yaml alone
 if (-Not (Test-Path "$observeagent_install_dir\observe-agent.yaml"))
